refactor: introduce domain ports and shared domain types
