+++ src/operator_exchange_component/README.md
@@ -1,19 +1,19 @@
-# Operator Exchange Component
+# Компонент внешнего обмена с эксплуатантами
 
-This component owns the trusted code that exchanges messages with external drone operators.
+Компонент содержит доверенный код, который обменивается сообщениями с внешними эксплуатантами дронов.
 
-Kafka and MQTT brokers are external infrastructure and are not part of this component's trusted
-code. The trusted code here is the adapter logic that:
+Kafka и MQTT-брокеры считаются внешней инфраструктурой и не входят в доверенный код компонента.
+В доверенную часть входит только логика адаптеров, которая:
 
-- publishes order requests to operators;
-- publishes confirmed prices to operators;
-- consumes `price_offer` messages;
-- consumes `order_result` messages;
-- updates order state through the persistence adapter.
+- публикует заявки эксплуатантам;
+- публикует подтвержденную цену эксплуатанту;
+- принимает сообщения `price_offer`;
+- принимает сообщения `order_result`;
+- обновляет состояние заказа через адаптер хранения.
 
-The component currently has Kafka and MQTT backends:
+Сейчас у компонента есть два backend-адаптера:
 
 - `kafka`
 - `mqtt`
 
-Both backends implement the same publishing behavior expected by the gateway publisher fan-out.
+Оба адаптера реализуют одинаковое поведение публикации, которое использует gateway publisher fan-out.
