+++ src/shared/store/README.md
@@ -1,0 +1,19 @@
+# PostgreSQL-адаптер хранения
+
+`shared/store` не является бизнес-компонентом и не должен рассматриваться как отдельный домен
+безопасности. Это инфраструктурный PostgreSQL-адаптер для текущей сборки агрегатора в одном
+Go-процессе. Доменные типы находятся в `shared/domain`; в `shared/store` оставлены type aliases
+для совместимости существующего кода.
+
+Доменное владение данными задается портами компонентов:
+
+- `registry_component.Store` - операции с заказчиками и эксплуатантами;
+- `orders_component.Store` - создание, чтение и поиск заказов;
+- `contracts_component.Store` - контрактные переходы состояния заказа;
+- `operator_exchange_component.Store` - изменения состояния заказа по сообщениям внешних эксплуатантов.
+
+Файл `ports_test.go` содержит проверки, что `*store.Store` реализует эти порты. Если метод случайно
+удалить или изменить несовместимо, `go test ./...` упадет.
+
+При дальнейшем выделении компонентов в отдельные сервисы этот пакет нужно будет разделить на
+отдельные persistence adapters или заменить реализации портов на клиенты межсервисного обмена.
